// src/services/models/albumResponse.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Swipe Pad API
 * OpenAPI spec version: 0.0.12
 */
import type { AlbumResponseType } from "./albumResponseType";
import type { AlbumResponseStatus } from "./albumResponseStatus";
import type { MemeResponse } from "./memeResponse";

export interface AlbumResponse {
  readonly walletAddress: string;
  readonly name: string;
  readonly type: AlbumResponseType;
  readonly status: AlbumResponseStatus;
  readonly image: string;
  readonly isTokenInAlbum: boolean;
  readonly memes: readonly MemeResponse[];
}